sllw, srlw updated; 21/90

+++ rv_emu.c
@@ -226,13 +226,12 @@ void emu_r_shift(struct rv_state_st *rsp, uint32_t iw) {
 	
 	if (funct3 == 0b001 && funct7 == 0b0000000) {
 		// sllw
-		rsp->regs[rd] = rsp->regs[rs1] << rsp->regs[rs2];
+		rsp->regs[rd] = (int32_t) (rsp->regs[rs1] << rsp->regs[rs2]);
 	} else if (funct3 == 0b101 && funct7 == 0b0000000) {
 		// srlw
-		rsp->regs[rd] = rsp->regs[rs1] >> rsp->regs[rs2];
+		rsp->regs[rd] = (int32_t) (rsp->regs[rs1] >> rsp->regs[rs2]);
 	} else if (funct3 == 0b101 && funct7 == 0b0100000) {
 		// sraw
-		// TODO test if cast works correctly
 		rsp->regs[rd] = (int64_t) (rsp->regs[rs1] >> rsp->regs[rs2]);
 	} else {
 		unsupported("R-type (shift word) funct3", funct3);
